Rename the Jet-H task for clarity
Rename the jet-hadron correlations task to clarify the purpose. In
particular, it previously contained initials, which could somewhat
confuse the user. Symlinks to the new task were created to ensure
that any previous references are not lost.
